+++ src/backend/mod.rs
@@ -60,9 +60,3 @@ pub mod serial;
     ))))
 )]
 pub mod vector;
-
-#[cfg(all(
-    target_os = "zkvm",
-    target_vendor = "succinct",
-))]
-pub mod zkvm;
