--- src/backend/zkvm/mod.rs
@@ -1,2 +1,0 @@
-mod edwards;
-pub mod scalar_mul;
