--- src/backend/zkvm/scalar_mul/mod.rs
@@ -1,5 +1,0 @@
-pub mod pippenger;
-pub mod precomputed_straus;
-pub mod straus;
-pub mod variable_base;
-pub mod vartime_double_base;
